+++ src/SmileiMPI/SmileiMPI_Cart2D.cpp
@@ -168,7 +168,7 @@ void SmileiMPI_Cart2D::createTopology(PicParams& params)
             params.n_space[i] =  params.nspace_win_x / number_of_procs[i];
             cell_starting_global_index[i] = coords_[i]*(params.nspace_win_x / number_of_procs[i]);
             
-            if ( number_of_procs[i]*params.n_space[i] != params.nspace_win_x ) {
+            if (number_of_procs[i]*params.n_space[i] != (unsigned int)params.nspace_win_x ) {
                 // Correction on the last MPI process of the direction to use the wished number of cells
                 if (coords_[i]==number_of_procs[i]-1) {
                     params.n_space[i] = params.nspace_win_x - params.n_space[i]*(number_of_procs[i]-1);
@@ -238,13 +238,13 @@ void SmileiMPI_Cart2D::exchangeParticles(Species* species, int ispec, PicParams&
         indexes_of_particles_to_exchange.clear();
         
         int tmp = 0;
-        for (int tid=0 ; tid < indexes_of_particles_to_exchange_per_thd->size() ; tid++)
+        for (int tid=0 ; tid < (int)indexes_of_particles_to_exchange_per_thd->size() ; tid++)
             tmp += ((*indexes_of_particles_to_exchange_per_thd)[tid]).size();
         indexes_of_particles_to_exchange.resize( tmp );
         
         int k=0;
-        for (int tid=0 ; tid < indexes_of_particles_to_exchange_per_thd->size() ; tid++) {
-            for (int ipart = 0 ; ipart < ((*indexes_of_particles_to_exchange_per_thd)[tid]).size() ; ipart++ ) {
+        for (int tid=0 ; tid < (int)indexes_of_particles_to_exchange_per_thd->size() ; tid++) {
+            for (int ipart = 0 ; ipart < (int) ((*indexes_of_particles_to_exchange_per_thd)[tid]).size() ; ipart++ ) {
                 indexes_of_particles_to_exchange[k] =  (*indexes_of_particles_to_exchange_per_thd)[tid][ipart] ;
                 k++;
             }
@@ -476,7 +476,7 @@ void SmileiMPI_Cart2D::exchangeParticles(Species* species, int ispec, PicParams&
         // Copy newly arrived particles back to the vector
         // WARNING: very different behaviour depending on which dimension particles are coming from.
         /********************************************************************************/
-        //We first evaluate how many particles arrive in each bin. 
+        //We first evaluate how many particles arrive in each bin.
 	if (iDim==1) {
 	    //1) Count particles coming from south and north
 	    for (int iNeighbor=0 ; iNeighbor<nbNeighbors_ ; iNeighbor++) {
@@ -537,6 +537,7 @@ void SmileiMPI_Cart2D::exchangeParticles(Species* species, int ispec, PicParams&
 		}
 	    }
 	}
+
         
         // Inject corner particles at the end of the list, update bmax
 	//if (iDim==cuParticles.dimension()-1) cout << "Number of diag particles " << diagonalParticles.size() << endl;
